opt code

+++ src/main.py
@@ -1,6 +1,7 @@
 import json
 import jieba
 import gensim
+import os
 
 
 def split_word(sentence):
@@ -27,31 +28,44 @@ with open(train_filepath, encoding='utf-8') as train_file:
 # 生成分词结果
 print("> 正在分词")
 content = []
-for value in data:
-    question = value['question']
-    content.append(split_word(question))
-with open(splitdata_filepath, 'w', encoding='utf-8') as f:
-    f.write(json.dumps(content, ensure_ascii=False))
+if os.path.exists(splitdata_filepath):
+    with open(splitdata_filepath, encoding='utf-8') as f:
+        content = json.load(f)
+else:
+    for value in data:
+        question = value['question']
+        content.append(split_word(question))
+    with open(splitdata_filepath, 'w', encoding='utf-8') as f:
+        f.write(json.dumps(content, ensure_ascii=False))
 
 
 # 生成gensim字典
 print("> 正在生成gensim字典")
-dictionary = gensim.corpora.Dictionary(content)
-dictionary.save(dictionary_filepath)
+if os.path.exists(dictionary_filepath):
+    dictionary = gensim.corpora.Dictionary.load(dictionary_filepath)
+else:
+    dictionary = gensim.corpora.Dictionary(content)
+    dictionary.save(dictionary_filepath)
 num_features = len(dictionary)
 corpus = [dictionary.doc2bow(line) for line in content]
 
 
 # 生成tfidf模型
-print("> 正在生成tfidf模型")
-tfidf = gensim.models.TfidfModel(corpus)
-tfidf.save(model_filepath)
+print("> 正在生成TF-IDF模型")
+if os.path.exists(model_filepath):
+    tfidf = gensim.models.TfidfModel.load(model_filepath)
+else:
+    tfidf = gensim.models.TfidfModel(corpus)
+    tfidf.save(model_filepath)
 
 
 # 生成tfidf相似度比较序列
-print("> 正在生成tfidf相似度比较序列")
-index = gensim.similarities.Similarity(index_filepath, tfidf[corpus], num_features)
-index.save(index_filepath)
+print("> 正在生成TF-IDF相似度比较序列")
+if os.path.exists(index_filepath):
+    index = gensim.similarities.Similarity.load(index_filepath)
+else:
+    index = gensim.similarities.Similarity(index_filepath, tfidf[corpus], num_features)
+    index.save(index_filepath)
 
 
 # 导入测试集进行测试
@@ -60,26 +74,28 @@ with open(test_filepath, encoding='utf-8') as test_file:
 print("> 正在使用测试集测试，共有" + str(len(test_data)) + "个问题")
 
 right = 0
-output_list = []
+output = []
 for value in test_data:
-    question = value['question']
-    sentences = split_word(question)     # 分词
-    vec = dictionary.doc2bow(sentences)  # 转词袋表示
+    question = value['question']         # 获得问题
+    sentence = split_word(question)      # 分词
+    vec = dictionary.doc2bow(sentence)   # 转词袋表示
     sims = index[tfidf[vec]]             # 相似度比较
     sorted_sims = sorted(enumerate(sims), key=lambda x: x[1], reverse=True)
-    i = sorted_sims[0][0]
+    i = sorted_sims[0][0]                # 最相似的问题的序号
 
-    output_list.append({'question:':value['question'], 'answer':data[i]['answer']}) 
+    # 输出问答对，若输出结果和参考结果不一致，则打印
+    output.append({'question:':value['question'], 'answer':data[i]['answer']}) 
     if(str(data[i]['answer']) == str(value['answer'])):
         right = right + 1
     else:
-        print("  答案不匹配：输出\""+str(data[i]['answer'])+"\"")
-        print("              参考\"" + str(value['answer']) + "\"")
+        print("  答案不匹配：输出结果\""+str(data[i]['answer'])+"\"")
+        print("              参考结果\"" + str(value['answer']) + "\"")
 
-with open(output_filepath, 'w', encoding='utf-8') as output:
-    output.write(json.dumps(output_list, ensure_ascii=False))
+with open(output_filepath, 'w', encoding='utf-8') as output_file:
+    output_file.write(json.dumps(output, ensure_ascii=False))
 
-print("  正确数量" + str(right) + " / " + str(len(test_data)))
+print("  正确数量：" + str(right) + " / " + str(len(test_data)))
+print("  正确率：" + str(right / len(test_data) * 100) + "%")
 
 
 # 允许用户继续输入问题
@@ -89,6 +105,6 @@ while True:
     vec = dictionary.doc2bow(sentences)  # 转词袋表示
     sims = index[tfidf[vec]]             # 相似度比较
     sorted_sims = sorted(enumerate(sims), key=lambda x: x[1], reverse=True)
-    i = sorted_sims[0][0]
+    i = sorted_sims[0][0]                # 最相似的问题的序号
     print(data[i]['answer'])
     sentences = input('Question: ')
